Adds style cop (#109)

+++ Projects/Scripts/Engines/Harvest/Core/HarvestVein.cs
@@ -2,7 +2,7 @@ namespace Server.Engines.Harvest
 {
   public class HarvestVein
   {
-    public HarvestVein(double veinChance, double chanceToFallback, HarvestResource primaryResource,
+    public HarvestVein(uint veinChance, double chanceToFallback, HarvestResource primaryResource,
       HarvestResource fallbackResource)
     {
       VeinChance = veinChance;
@@ -11,12 +11,12 @@ namespace Server.Engines.Harvest
       FallbackResource = fallbackResource;
     }
 
-    public double VeinChance{ get; set; }
+    public uint VeinChance { get; set; }
 
-    public double ChanceToFallback{ get; set; }
+    public double ChanceToFallback { get; set; }
 
-    public HarvestResource PrimaryResource{ get; set; }
+    public HarvestResource PrimaryResource { get; set; }
 
-    public HarvestResource FallbackResource{ get; set; }
+    public HarvestResource FallbackResource { get; set; }
   }
 }
